Update Bulb Switcher (319).java

+++ Backtracking/Game/Bulb Switcher (319).java
@@ -1,5 +1,5 @@
 public class Solution {
     public int bulbSwitch(int n) {
-        return Math.sqrt(n);
+        return (int) Math.sqrt(n);
     }
 }
